+++ js/main.js
@@ -1,9 +1,7 @@
 import showPictures from './modules/show-pictures.js';
 import addPicture from './modules/add-picture.js';
-import photoScale from './modules/photo-scale.js';
-import photoEffects from './modules/photo-effects.js';
+//import photoEffects from './modules/photo-effects.js';
 
 showPictures();
 addPicture();
-photoScale();
-photoEffects();
+//photoEffects();
