+++ js/modules/add-picture.js
@@ -1,8 +1,5 @@
-import {testStringLength} from './test-string.js';
-
-const MAX_COMMENT_LENGTH = 140;
-const MAX_HASH_LENGTH = 20;
-const MAX_HASH_ARRAY_LENGTH = 5;
+import {checkCommentValidity, checkHashValidity} from './check-validity.js';
+import {scaleEffect} from './photo-effects.js';
 
 export default function addPicture() {
   const body = document.querySelector('body');
@@ -11,62 +8,10 @@ export default function addPicture() {
   const imageUpload = uploadForm.querySelector('.img-upload__overlay');
   const imageUploadCancel = uploadForm.querySelector('.img-upload__cancel');
   const textHashtags = uploadForm.querySelector('.text__hashtags');
-  const textDescription = uploadForm.querySelector('.text__description');
+  const textDescription = document.querySelector('.text__description');
   const imgPreview = uploadForm.querySelector('.img-upload__preview-img');
-  const commentReg = /[~`!@_()$%^&*+=\-[\]\\';,/{}|\\":<>?]/g;
-
-  const hasDuplicates = (array) => (new Set(array)).size !== array.length;
-
-  const checkCommentValidity = () => {
-    if (!testStringLength(MAX_COMMENT_LENGTH, textDescription.value)) {
-      textDescription.setCustomValidity('до 140 символов');
-    } else {
-      textDescription.setCustomValidity('');
-    }
-    textDescription.reportValidity();
-  };
-
-  const checkHashValidity = () => {
-    textHashtags.value = textHashtags.value.replace(/\s+/g, ' ');
-    const hashArray = textHashtags.value.toLowerCase().split(' ');
-    let error = '';
-
-    hashArray.forEach((hash) => {
-      if(!hash.startsWith('#')) {
-        error = 'хеш-тег должен начинаться с решётки #';
-      }
-
-      if(hash === '#') {
-        error = 'хеш-тег не может состоять только из одной решётки';
-      }
-
-      if(commentReg.test(hash)) {
-        error = 'хештег не может содержать пробелы, спецсимволы (@, $ и т. п.)';
-      }
-
-      if(hash.length > MAX_HASH_LENGTH) {
-        error = 'не больше 20 символов';
-      }
-
-      if(hasDuplicates(hashArray)) {
-        error = 'хеш-теги не могут повторяться';
-      }
-
-      if (hashArray.length > MAX_HASH_ARRAY_LENGTH) {
-        error = 'не больше 5 тегов';
-      }
-    });
-
-    if (!error) {
-      textHashtags.setCustomValidity('');
-    } else if (hashArray[0] === '') {
-      textHashtags.value = textHashtags.value.trim();
-    } else {
-      textHashtags.setCustomValidity(error);
-    }
-
-    textHashtags.reportValidity();
-  };
+  const scaleButton = document.querySelector('.img-upload__scale');
+  const scaleInput = document.querySelector('.scale__control--value');
 
   const onUploadEscKeydown = (evt) => {
     if (!evt.target.closest('.img-upload__text') && evt.key === 'Escape' ) {
@@ -85,27 +30,29 @@ export default function addPicture() {
 
     document.addEventListener('keydown', onUploadEscKeydown);
     imageUploadCancel.addEventListener('click', onCancelClick);
-
     textDescription.addEventListener('input', checkCommentValidity);
     textHashtags.addEventListener('input', checkHashValidity);
+    //scale effect
+    scaleButton.addEventListener('click', scaleEffect);
+    scaleInput.value = 100;
+    imgPreview.style.transform = 'scale(1.0)';
   }
 
   function uploadModalClose () {
     imageUpload.classList.add('hidden');
     body.classList.remove('modal-open');
     uploadForm.reset();
+    uploadFile.value = '';
 
     document.removeEventListener('keydown', onUploadEscKeydown);
     imageUploadCancel.removeEventListener('click', onCancelClick);
-
     textDescription.removeEventListener('input', checkCommentValidity);
     textHashtags.removeEventListener('input', checkHashValidity);
-    uploadFile.value = '';
+    scaleButton.removeEventListener('click', scaleEffect);
   }
 
   uploadFile.addEventListener('change', () => {
     uploadModalOpen();
     imgPreview.src = URL.createObjectURL(uploadFile.files[0]);
   });
-
 }
